% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_images_and_labels.R
\name{add_categorical_species}
\alias{add_categorical_species}
\title{Adds categorical variables to each image}
\usage{
add_categorical_species(cfg, image, species)
}
\arguments{
\item{cfg}{list defining psptools configuration}

\item{image}{2D array where each row is an image and the columns are toxins and environmentals from each week}

\item{species}{vector of species the length of `image`}
}
\value{
image with additional cols representing categorical species
}
\description{
Adds categorical variables to each image
}
