% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_functions.R
\name{balance_classes}
\alias{balance_classes}
\title{Function to balaance training data so all classes are distributed evenly (downsample)}
\usage{
balance_classes(s)
}
\arguments{
\item{s}{a list of images with imbalanced class frequencies}
}
\value{
list of images with equal class frequencies
}
\description{
Function to balaance training data so all classes are distributed evenly (downsample)
}
