% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.R
\name{check_gap}
\alias{check_gap}
\title{Checks gap_days column to filter for only}
\usage{
check_gap(x, minimum_gap, maximum_gap)
}
\arguments{
\item{x}{gap days column from raw data}

\item{minimum_gap}{gaps allowed into an image must be greater than}

\item{maximum_gap}{gaps allowed into an image must be less than}
}
\value{
logical for each row in data - TRUE if gap meets gap criteria, FALSE if not
}
\description{
Checks gap_days column to filter for only
}
