% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_functions.R
\name{compute_batches}
\alias{compute_batches}
\title{For a given station and year, find the indices of each batch (image) that can be made from the subset}
\usage{
compute_batches(nbatches, steps)
}
\arguments{
\item{nbatches}{number of batches that can be made for a given year and station}

\item{steps}{number of weeks in the image}
}
\value{
batches
}
\description{
For a given station and year, find the indices of each batch (image) that can be made from the subset
}
