% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.R
\name{compute_gap}
\alias{compute_gap}
\title{Finds gap between samples in the same season at the same site}
\usage{
compute_gap(x)
}
\arguments{
\item{x}{tibble of toxin measurement data}
}
\value{
the input tibble with a column `gap_days` containing the difference between samples in the same season
}
\description{
Finds gap between samples in the same season at the same site
}
