% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_functions.R
\name{forecast_dummy}
\alias{forecast_dummy}
\title{Adds a dummy row for forecast targets (which have no data in our package, but need to meet the gap filtering when making images)}
\usage{
forecast_dummy(data)
}
\arguments{
\item{data}{the raw data before going into `make_image_list()`}
}
\value{
a modified tibble where each site with data this year has an additional dummy row to make an image for a forecast
}
\description{
Adds a dummy row for forecast targets (which have no data in our package, but need to meet the gap filtering when making images)
}
