% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_metrics.R
\name{forecast_metrics}
\alias{forecast_metrics}
\title{Finds metrics from a table of forecast predictions (usually the output of `make_forecast_list()`)}
\usage{
forecast_metrics(
  fc,
  predicted_col = "predicted_class",
  measured_col = "actual_class",
  closure_class = 3
)
}
\arguments{
\item{fc}{tibble of predictions}

\item{predicted_col}{character string of predicted classification column name}

\item{measured_col}{character string of measured classification column}

\item{closure_class}{integer defining closure-level class (usually 3, but 1 in case of binary predictions)}
}
\description{
Finds metrics from a table of forecast predictions (usually the output of `make_forecast_list()`)
}
