% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_model.R
\name{forecast_model}
\alias{forecast_model}
\title{Function to build, compile and fit the model to the training data.
The model is then evaluates with the test data.}
\usage{
forecast_model(cfg, model_input, forecast_mode = FALSE)
}
\arguments{
\item{cfg}{model configuration file}

\item{model_input}{output of transform_data()}

\item{forecast_mode}{logical if testing a model with unknown test targets (aka making a forecast with a trained model)}
}
\value{
forecast object containing an evaluation of the model, information about test data and actual classifications
\itemize{
\item{forecast list containing prediction information along with model object and history. if forecast_mode==FALSE, also returns test set information and metrics}
\item{model fitted keras model object}
\item{history a keras_training_history object containing metrics from model fitting}}
}
\description{
Function to build, compile and fit the model to the training data.
The model is then evaluates with the test data.
}
