% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_functions.R
\name{get_class_weights}
\alias{get_class_weights}
\title{Find class weights for imbalanced model input (training data)}
\usage{
get_class_weights(train_classes)
}
\arguments{
\item{train_classes}{labels for training data (target toxicity classification)}
}
\value{
class weight named list
}
\description{
Find class weights for imbalanced model input (training data)
}
