% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{is_cfg_valid}
\alias{is_cfg_valid}
\title{Tests if required configuration options are present}
\usage{
is_cfg_valid(cfg)
}
\arguments{
\item{cfg}{a psptools configuration}
}
\value{
NULL if pass or error if fail
}
\description{
Tests if required configuration options are present
}
