% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{load_and_predict}
\alias{load_and_predict}
\title{Loads a pre-trained forecast model and makes predictions}
\usage{
load_and_predict(cfg, model_input, model_dir, model_name)
}
\arguments{
\item{cfg}{a list containing model configuration arguments}

\item{model_input}{a model_input object containing lists for train and test sets}

\item{model_dir}{character path to directory containing pre-trained models}

\item{model_name}{character name of model to load}
}
\value{
tibble of predictions in forecast_list format
}
\description{
Loads a pre-trained forecast model and makes predictions
}
