% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_io.R
\name{load_psp_model}
\alias{load_psp_model}
\title{Loads a pre-trained PSP forecast model from the saved_models directory}
\usage{
load_psp_model(model_name)
}
\arguments{
\item{model_name}{character name of saved model}
}
\description{
Loads a pre-trained PSP forecast model from the saved_models directory
}
