% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_functions.R
\name{log_inputs}
\alias{log_inputs}
\title{Function to log a list of inputs - should this log all of them?}
\usage{
log_inputs(data, cfg)
}
\arguments{
\item{data}{raw input data}

\item{cfg}{list of model configuration parameters}
}
\value{
same data as input but with log transformed variables
}
\description{
Function to log a list of inputs - should this log all of them?
}
