% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_confusion_matrix.R
\name{make_confusion_matrix}
\alias{make_confusion_matrix}
\title{Function to plot confusion matrix from forecast model test run with past data (known targets in test data)}
\usage{
make_confusion_matrix(cfg, forecast_list)
}
\arguments{
\item{cfg}{the configuration object used to build and train the model}

\item{forecast_list}{formatted predictions from forecast model}
}
\value{
confusion_matrix plot of actual vs predicted labels in the test set
}
\description{
Function to plot confusion matrix from forecast model test run with past data (known targets in test data)
}
