% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forecast_list.R
\name{make_forecast_list}
\alias{make_forecast_list}
\title{Formats the output from the PSP forecast model in standard form to be added to forecast database}
\usage{
make_forecast_list(cfg, model, forecast_mode = FALSE)
}
\arguments{
\item{cfg}{the configuration object used to build and train the model}

\item{model}{a list of predictions made by the trained neural network and their metadata}

\item{forecast_mode}{logical if testing a model with unknown test targets (aka making a forecast with a trained model)}
}
\value{
a list of forecast classifications with their metadata
}
\description{
Formats the output from the PSP forecast model in standard form to be added to forecast database
}
