% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_functions.R
\name{n_batches}
\alias{n_batches}
\title{For a given station and year, find how many batches(images) can be made}
\usage{
n_batches(nx, steps)
}
\arguments{
\item{nx}{number of samples for that year and station}

\item{steps}{number of weeks in the image}
}
\value{
nbatches number of batches that can be made for a given year and station
}
\description{
For a given station and year, find how many batches(images) can be made
}
