% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_functions.R
\name{normalize_input}
\alias{normalize_input}
\title{Takes toxin and environmental input columns and normalizes to range 0-1}
\usage{
normalize_input(x, toxins, environmentals)
}
\arguments{
\item{x}{tibble of raw data}

\item{toxins}{toxins in raw data}

\item{environmentals}{environmental variables}
}
\value{
tibble of normalized input columns
}
\description{
Takes toxin and environmental input columns and normalizes to range 0-1
}
