% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_images_and_labels.R
\name{pool_images_and_labels}
\alias{pool_images_and_labels}
\title{Creates images and labels for input into neural net
Image takes all images in image_list and stretches them into an array
Labels takes classifications and categorizes them for tensorflow input}
\usage{
pool_images_and_labels(
  image_list_subset,
  cfg,
  missing_value = 0.5,
  downsample = FALSE,
  upsample = FALSE
)
}
\arguments{
\item{image_list_subset}{subset of image list from make_image_list() for either training or testing data}

\item{cfg}{model configuration file}

\item{missing_value}{value to replace na with}

\item{downsample}{logical indicating whether or not to balance the frequency of each class in the training images}

\item{upsample}{logical to call a function that balances class distribution by upsampling rare classes}
}
\value{
list containing the formatted images as tensorflow model input (2D array) and vectors of other metadata.
 only non-forecast/test images will have categorical labels, classifications and toxicity for image outcome.
\itemize{
\item{image a 2 dimensional array where each row is an image and the columns are input variables from each week}
\item{locations location_ids for each image}
\item{dates dates of the final week of each image}
\item{species species for each image}
\item{labels categorical label for each image based on classification}
\item{classifications toxicity classification of each image}
\item{toxicity raw PSP toxicity value for each image}
}
}
\description{
Creates images and labels for input into neural net
Image takes all images in image_list and stretches them into an array
Labels takes classifications and categorizes them for tensorflow input
}
