% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Function to read configuration yaml file for model}
\usage{
read_config(filename, autopopulate = "num_classes", old = FALSE)
}
\arguments{
\item{filename}{name of config yaml file}

\item{autopopulate}{optional argument of fields to autopopulate}

\item{old}{logical if true then config file will be read from `configurations/old`}
}
\value{
config list of configuration elements for model
}
\description{
Function to read configuration yaml file for model
}
