% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_images_and_labels.R
\name{replace_na}
\alias{replace_na}
\title{Replace any NA values with specified missing value}
\usage{
replace_na(x, missing_value = -1)
}
\arguments{
\item{x}{2D array image object}

\item{missing_value}{integer to replace NA toxin levels with}
}
\value{
x
}
\description{
Replace any NA values with specified missing value
}
