% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ensemble_test.R
\name{run_ensemble_test}
\alias{run_ensemble_test}
\title{Runs a model configuration n times and produces metrics/plots of results}
\usage{
run_ensemble_test(cfg, input_data, n_runs, write_results = FALSE)
}
\arguments{
\item{cfg}{a list containing configuration params}

\item{input_data}{a tibble of raw input data (psp measurements, environmental data)}

\item{n_runs}{integer of times to run the model for the test}

\item{write_results}{logical if true then results/metrics get appended to results files}
}
\value{
a list containing predictions, metrics and plots
}
\description{
Runs a model configuration n times and produces metrics/plots of results
}
