% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Runs the multiclass script - can be used for testing multiple iterations of the model}
\usage{
run_model(cfg, model_input)
}
\arguments{
\item{cfg}{a list containing model configuration arguments}

\item{model_input}{a model_input object containing lists for train and test sets}
}
\value{
a one row tibble containing metrics of model run
}
\description{
Runs the multiclass script - can be used for testing multiple iterations of the model
}
