% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_io.R
\name{save_psp_model}
\alias{save_psp_model}
\title{Saves a pre-trained PSP forecast model to the saved_models directory}
\usage{
save_psp_model(model, model_name)
}
\arguments{
\item{model}{a pre-trained model}

\item{model_name}{character name to save model as}
}
\description{
Saves a pre-trained PSP forecast model to the saved_models directory
}
