% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_functions.R
\name{upsample}
\alias{upsample}
\title{Balances class distribution in training set by adding duplicates to classes that are more rare}
\usage{
upsample(images)
}
\arguments{
\item{images}{a list of images with an imbalanced class distribution}
}
\value{
a training set of images with each class appearing the same number of times
}
\description{
Balances class distribution in training set by adding duplicates to classes that are more rare
}
