% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_functions.R
\name{validation_splitter}
\alias{validation_splitter}
\title{Function to generate a validation set of labeled samples with an even distribution of each class}
\usage{
validation_splitter(image_list_subset, cfg)
}
\arguments{
\item{image_list_subset}{subset of images for training}

\item{cfg}{configuration file for model}
}
\value{
list of images to use as a validation set
}
\description{
Function to generate a validation set of labeled samples with an even distribution of each class
}
