% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ensemble_forecast.R
\name{make_ensemble_forecast}
\alias{make_ensemble_forecast}
\title{Transforms predictions from ensemble model test runs and combines them into an ensemble forecast with metrics}
\usage{
make_ensemble_forecast(cfg, predictions)
}
\arguments{
\item{cfg}{a configuration list}

\item{predictions}{a table of predictions where each site-date has multiple entries}
}
\value{
a list containing a table of predictions where each site-date has one prediction and metrics
}
\description{
Transforms predictions from ensemble model test runs and combines them into an ensemble forecast with metrics
}
