% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_image_list.R
\name{make_image_list}
\alias{make_image_list}
\title{Takes raw input data, filters for image criteria and creates images with dimensions (n_steps + forecast steps) x length(toxins + environmentals)}
\usage{
make_image_list(input_data, cfg)
}
\arguments{
\item{input_data}{tibble containing PSP measurements with their date sampled, location, shellfish species and additional covariates}

\item{cfg}{psptools configuration}
}
\value{
list of lists containing an image along with its associated data (location_id, date, etc.)
\itemize{
\item{status logical if the image passes the image gap criteria (gap >= minimum_gap & gap <= maximum_gap)}
\item{year the year the image is from}
\item{location_id the sampling station id}
\item{species the shellfish species the image was made for}
\item{date the date of the final row in the image (the forecast is for forecast_steps ahead of this date)}
\item{toxixty the total toxicity used to regress on instead of classify a binned toxicity}
\item{classification the classification (0:num_classes) of the final row in the image}
\item{image a 2 dimensional array with the dimensions (n_steps + forecast steps) x length(toxins + environmentals)}
}
}
\description{
Takes raw input data, filters for image criteria and creates images with dimensions (n_steps + forecast steps) x length(toxins + environmentals)
}
