% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_images_and_labels.R
\name{pool_images_and_labels}
\alias{pool_images_and_labels}
\title{Creates images and labels for input into neural net
Image takes all images in psp_lst and stretches them into an array
Labels takes classifications and categorizes them for nn input}
\usage{
pool_images_and_labels(
  image_list_subset,
  cfg,
  missing_value = 0.5,
  scaling_factors = NULL,
  scaling = c("normalize", "input_scale")[2],
  downsample = FALSE,
  upsample = FALSE
)
}
\arguments{
\item{image_list_subset}{subset of image list from make_image_list() for either training or testing data}

\item{cfg}{model configuration file}

\item{missing_value}{value to replace na with}

\item{scaling_factors}{null if training data; training data scaling factors are passed to scale testing data}

\item{scaling}{selected method to scale input data}

\item{downsample}{logical indicating whether or not to balance the frequency of each class in the training images}

\item{upsample}{logical to call a function that balances class distribution by upsampling rare classes}
}
\value{
list containing the formatted images and their labels as keras model input, and additional data
\itemize{
\item{labels}
\item{image a 2 dimensional array where each row is an image and the columns are toxins and environmentals from each week}
\item{classifications the classification of each image}
\item{locations the sampling station of each image}
\item{dates the date of the final week of each image}
\item{scaling_factors}
}
}
\description{
Creates images and labels for input into neural net
Image takes all images in psp_lst and stretches them into an array
Labels takes classifications and categorizes them for nn input
}
