% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{train_and_predict}
\alias{train_and_predict}
\title{Trains a forecast model and makes predictions}
\usage{
train_and_predict(cfg, model_input, forecast_mode)
}
\arguments{
\item{cfg}{a list containing model configuration arguments}

\item{model_input}{a model_input object containing lists for train and test sets}

\item{forecast_mode}{logical to define if testing on historical data or predicting in real-time}
}
\value{
a tibble of predictions in forecast_list format along with metrics if testing
}
\description{
Trains a forecast model and makes predictions
}
